backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
celery==5.3.6
redis==5.0.1
drf-spectacular==0.27.0
razorpay==1.4.1